TCDynamics/requirements.txt
# Azure Functions Python Dependencies
azure-functions==1.18.0
azure-functions-worker==1.0.0
azure-storage-blob==12.19.0
azure-cosmos==4.5.1
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
email-validator==2.1.0

